+++ library/xgpl_src/supplementary_driver/driver/inputs/icounteri_c.h
@@ -102,7 +102,7 @@ public:
     @return frequency calculated from time between last two signals
             or 0 if time is longer than initially given timebase
   */
-  uint16_t frequency(){return CounterI_c::frequency();}
+  uint32_t frequency(){return CounterI_c::frequency();}
 
   /**
    get time since last signal
