+++ library/xgpl_src/supplementary_driver/hal/esxu/inputs/inputs_target_extensions.h
@@ -69,7 +69,7 @@ uint16_t getCounterPeriod(uint8_t ab_channel);
   @return frequency calculated from time between last two signals
           or 0 if time is longer than initially given timebase
 */
-uint16_t getCounterFrequency(uint8_t ab_channel);
+uint32_t getCounterFrequency(uint8_t ab_channel);
 /**
  get time since last signal and reset according trigger timers
  if timebase is exceeded -> avoid overflow problems if timer floated around 0xFFFF
